+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Remove repository card shortlink ([#137](https://github.com/scm-manager/scm-review-plugin/pull/137))
+
 ### Fixed
 - Close toast notification after refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))
 - Update comment content on refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))
+
 
 ## 2.9.1 - 2021-05-05
 ### Changed