+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 - Use react-query to enable frontend caching ([#138](https://github.com/scm-manager/scm-review-plugin/pull/138))
 
+## 2.9.2 - 2021-07-06
 ### Fixed
 - Close toast notification after refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))
 - Update comment content on refresh ([#136](https://github.com/scm-manager/scm-review-plugin/pull/136))